+++ src/main/kotlin/com/melvic/archia/ast/compound/DisMaxQuery.kt
@@ -4,7 +4,7 @@ import com.melvic.archia.ast.*
 
 data class DisMaxQuery(
     var _queries: MultiClause? = null,
-    var tieBreaker: Float? = null
+    var tieBreaker: Double? = null
 ): Clause {
     fun queries(init: Init<ClauseArrayBuilder>) {
         _queries = ClauseArrayBuilder().apply(init).clauses
